Weight Attribute
A default weight value of 100(g) has been set in the AddMealActivity.
Food objects created in the AddMealActivity will now be given a weight instead of defaulting to 0g.

Profile now has a toJSON() to store target calorie, carb, protein, fat in a local database.

**minor**
ArrayAdapter is now given a Meal's menu object as opposed to a pointer to Meal's menu object.