Main Controller Class
A MainController class has been created. This class is used as a "Global" variable to provide access to the same Profile object in every activity.

SummaryFragment's MealList populates directly from a Profile object's Menu object.

ReportsFragment now populates directly from a Profile object's Meal.

AddMealActivity creats a Meal and adds it to Profile. Any food created in this activity is added to that cooresponding Meal.